Desenvolvendo os últimos dois serviços - serviço que retorna funcionários agrupados por UF de Nascimento, de forma quantitativa e serviço que localiza/retorna funcionários por faixa salarial.

+++ FuncionarioWS/src/java/model/mdlFuncionario.java
@@ -16,7 +16,7 @@ public class mdlFuncionario {
     private String cpf;
     private String nome;
     private String ufnasc;
-    private Float salario;
+    private Double salario;
     private String status;
 
     public String getDataCad() {
@@ -39,7 +39,7 @@ public class mdlFuncionario {
         return ufnasc;
     }
     
-    public Float getSalario() {
+    public Double getSalario() {
         return salario;
     }
     
@@ -67,7 +67,7 @@ public class mdlFuncionario {
         this.ufnasc = ufnasc;
     }
     
-    public void setSalario(Float salario) {
+    public void setSalario(Double salario) {
         this.salario = salario;
     }
